Move dictionary files to subdirectory
