+++ src/web/client/root.tsx
@@ -11,7 +11,7 @@ import ThemeProvider from "@mui/material/styles/ThemeProvider";
 import { Macronizer } from "@/web/client/pages/macron";
 import { SinglePageApp } from "@/web/client/components/single_page_app";
 import { About } from "@/web/client/pages/about";
-import { DictionaryView } from "@/web/client/pages/dictionary";
+import { DictionaryView } from "@/web/client/pages/dictionary/dictionary";
 import { Solarized } from "@/web/client/colors";
 import { Router } from "@/web/client/components/router";
 import { SettingsHandler } from "@/web/client/components/global_flags";
